+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED 2.1.0] - 2019-12-06
+## Removed
+- Removed `bundles/cms/js/file-picker.js`
+## Changed
+- Rewrote `admin-backend.js` and `admin-frontend.js
+## Todo
+- Remove `bundles/cms/js/admin-backend_alt.js` from config.yaml
+
 ## [2.0.20] - 2019-12-04
 ## Changed
 - Don't include pages in sitemap that have a `noindex` in `page.meta.robots`